combined ninject modules
